# service/apigateway/docs/swagger.yaml
definitions:
  errors.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
      retryable:
        type: boolean
      status:
        type: string
      trace_id:
        type: string
      type:
        $ref: '#/definitions/errors.ErrorType'
      validations:
        items:
          $ref: '#/definitions/errors.ValidationError'
        type: array
    type: object
  errors.ErrorType:
    enum:
    - INTERNAL
    - NOT_FOUND
    - BAD_REQUEST
    - CONFLICT
    - UNAUTHORIZED
    - FORBIDDEN
    - TIMEOUT
    - UNAVAILABLE
    type: string
    x-enum-varnames:
    - ErrorTypeInternal
    - ErrorTypeNotFound
    - ErrorTypeBadRequest
    - ErrorTypeConflict
    - ErrorTypeUnauthorized
    - ErrorTypeForbidden
    - ErrorTypeTimeout
    - ErrorTypeUnavailable
  errors.ValidationError:
    properties:
      field:
        type: string
      message:
        type: string
    type: object
...
  response.ApiResponseBanner:
    properties:
      data:
        $ref: '#/definitions/response.BannerResponse'
      message:
        type: string
      status:
        type: string
    type: object
...
  response.ApiResponsePaginationBanner:
    properties:
      data:
        items:
          $ref: '#/definitions/response.BannerResponse'
        type: array
      message:
        type: string
      pagination:
        $ref: '#/definitions/response.PaginationMeta'
      status:
        type: string
    type: object
  response.ApiResponsePaginationBannerDeleteAt:
    properties:
      data:
        items:
          $ref: '#/definitions/response.BannerResponseDeleteAt'
        type: array
      message:
        type: string
      pagination:
        $ref: '#/definitions/response.PaginationMeta'
      status:
        type: string
    type: object
...
  response.BannerResponse:
    properties:
      created_at:
        type: string
      end_date:
        type: string
      end_time:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
      start_date:
        type: string
      start_time:
        type: string
      updated_at:
        type: string
    type: object
  response.BannerResponseDeleteAt:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      end_date:
        type: string
      end_time:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
      start_date:
        type: string
      start_time:
        type: string
      updated_at:
        type: string
    type: object
...
  response.PaginationMeta:
    properties:
      current_page:
        type: integer
      page_size:
        type: integer
      total_pages:
        type: integer
      total_records:
        type: integer
    type: object
...
info:
  contact: {}
paths:
...
  /api/banner-query:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all banners
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: page_size
        type: integer
      - description: Search query
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of banners
          schema:
            $ref: '#/definitions/response.ApiResponsePaginationBanner'
        "500":
          description: Failed to retrieve banner data
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - Bearer: []
      summary: Find all banners
      tags:
      - Banner Query
  /api/banner-query/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve a single banner by its ID
      parameters:
      - description: Banner ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Banner data
          schema:
            $ref: '#/definitions/response.ApiResponseBanner'
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Failed to retrieve banner data
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - Bearer: []
      summary: Find a banner by ID
      tags:
      - Banner Query
  /api/banner-query/active:
    get:
      consumes:
      - application/json
      description: Retrieve a list of active banners
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: page_size
        type: integer
      - description: Search query
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of active banners
          schema:
            $ref: '#/definitions/response.ApiResponsePaginationBannerDeleteAt'
        "500":
          description: Failed to retrieve banner data
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - Bearer: []
      summary: Find active banners
      tags:
      - Banner Query
  /api/banner-query/trashed:
    get:
      consumes:
      - application/json
      description: Retrieve a list of trashed banners
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: page_size
        type: integer
      - description: Search query
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of trashed banners
          schema:
            $ref: '#/definitions/response.ApiResponsePaginationBannerDeleteAt'
        "500":
          description: Failed to retrieve banner data
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - Bearer: []
      summary: Find trashed banners
      tags:
      - Banner Query
...
swagger: "2.0"